새로운 모델 Object 디자인 추가

+++ architecture-common/src/main/java/architecture/common/model/v2/ModelObjectTypeManager.java
@@ -29,6 +29,6 @@ public interface ModelObjectTypeManager {
 	
 	public abstract ModelObjectType getModelObjectType( String code );
 	
-	public abstract ModelObjectType getModelObjectType( int objectId );
+	public abstract ModelObjectType getModelObjectType( Long objectId );
 	
 }
